+++ be_atom_triple_/index.js
@@ -1,38 +1,28 @@
-import { be_, be_arg_triple__new } from 'ctx-core/be'
+import { be_ } from 'ctx-core/be'
 import { atom_ } from '../atom_/index.js'
 /** @typedef {import('@ctx-core/object').be__params_T}be__params_T */
 /** @typedef {import('@ctx-core/object').Ctx}Ctx */
 /** @typedef {import('nanostores').WritableAtom}WritableAtom */
 /** @typedef {import('./index.d.ts').be_atom_triple_T}be_atom_triple_T */
 /**
+ * @param {be__val__new_T<unknown>}val__new
  * @returns {be_atom_triple_T}
  * @private
  */
-export function be_atom_triple_(...arg_a) {
-	let [
-		id,
-		atom__new,
-		be__params
-	] = be_arg_triple__new(...arg_a)
-	if (!atom__new) {
-		atom__new =
-			()=>
-				atom_()
-	}
-	const _be_ =
-		(be__params && be__params.be_)
-		?? be_
-	const val$_ =
-		id
-		? _be_(id, atom__new, be__params)
-		: _be_(atom__new, be__params)
-	const val_ = ctx=>val$_(ctx)()
-	const val__set = (ctx, val)=>{
-		val$_(ctx).set(val)
-	}
-	return [
-		val$_,
-		val_,
-		val__set
+export function be_atom_triple_(val__new) {
+	let oninit
+	const be_atom_triple = [
+		be_(ctx=>{
+			let atom = atom_(val__new(ctx))
+			oninit?.(atom)
+			return atom
+		}),
+		ctx=>be_atom_triple[0](ctx)(),
+		(ctx, val)=>{
+			be_atom_triple[0](ctx).set(val)
+		},
 	]
+	be_atom_triple.config = config__fn=>(config__fn(be_atom_triple[0]), be_atom_triple)
+	be_atom_triple.oninit__def = _oninit=>(oninit = _oninit, be_atom_triple)
+	return be_atom_triple
 }