+++ be_computed_pair_/index.js
@@ -1,27 +1,23 @@
-import { be_, be_arg_triple__new } from 'ctx-core/be'
-/** @typedef {import('@ctx-core/object').be__params_T}be__params_T */
-/** @typedef {import('@ctx-core/object').Ctx}Ctx */
-/** @typedef {import('nanostores').ReadableAtom}ReadableAtom */
-/** @typedef {import('./index.d.ts').be_computed_pair_T}be_computed_pair_T */
+import { be_ } from 'ctx-core/be'
+import { computed_ } from '../computed_/index.js'
+/** @typedef {import('@ctx-core/object').be__val__new_T} */
+/** @typedef {import('./index.d.ts').be_computed_pair_T} */
 /**
+ * @param {be__val__new_T<unknown>}val__new
  * @returns {be_computed_pair_T}
  * @private
  */
-export function be_computed_pair_(...arg_a) {
-	const [id, computed__new, be__params] =
-		/** @type {[string, (ctx:Ctx)=>ReadableAtom, be__params_T]} */
-		be_arg_triple__new(...arg_a)
-	if (!computed__new) throw new Error('be_computed_pair_|computed__new argument is required')
-	const _be_ =
-		(be__params && be__params.be_)
-		?? be_
-	const val$_ =
-		id
-		? _be_(id, computed__new, be__params)
-		: _be_(computed__new, be__params)
-	const val_ = ctx=>val$_(ctx)()
-	return [
-		val$_,
-		val_,
+export function be_computed_pair_(val__new) {
+	let oninit
+	const be_computed_pair = [
+		be_(ctx=>{
+			let computed$ = computed_(()=>val__new(ctx))
+			oninit?.(computed$)
+			return computed$
+		}),
+		ctx=>be_computed_pair[0](ctx)(),
 	]
+	be_computed_pair.config = config__fn=>(config__fn(be_computed_pair[0]), be_computed_pair)
+	be_computed_pair.oninit__def = _oninit=>(oninit = _oninit, be_computed_pair)
+	return be_computed_pair
 }